+++ src/main/java/frc/robot/commands/Gripper/ReleaseCommand.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Gripper;
 
 import edu.wpi.first.wpilibj2.command.Command;
 import frc.robot.subsystems.Gripper.Gripper;
@@ -34,7 +34,7 @@ public class ReleaseCommand extends Command {
 
         gripper.setPercentOutput(GripperConstants.CORAL_EJECT_POWER);
         gripper.setGamePiece(GamePiece.None);
-        
+
     }
   }
 }
